+++ src/low_level/servodriver.cpp
@@ -1,7 +1,7 @@
 #include "low_level/servodriver.h"
 #include <Arduino.h>
 
-Servo::Servo(uint8_t m_pwmPin, float m_minDutyCycle, float m_maxDutyCycle) {
+Servo::Servo(uint8_t m_pwmPin, double m_minDutyCycle, double m_maxDutyCycle) {
   pwmPin = m_pwmPin;
   minDutyCycle = m_minDutyCycle * 0.01;
   maxDutyCycle = m_maxDutyCycle * 0.01;
@@ -9,11 +9,11 @@ Servo::Servo(uint8_t m_pwmPin, float m_minDutyCycle, float m_maxDutyCycle) {
 
 void Servo::init(void) { ledcAttach(pwmPin, 50, 12); }
 
-void Servo::setAngle(float angle) {
+void Servo::setAngle(double angle) {
   angle = angle > 180 ? 180 : angle;
   angle = angle < 0 ? 0 : angle;
 
-  float dutyCycle =
+  double dutyCycle =
       minDutyCycle + (angle / 180.0) * (maxDutyCycle - minDutyCycle);
